chore: move rust-specific tidying to the rust package
The Rust-specific code is within the rust package to match other
languages, and to make the top-level tidy code more
language-agnostic. One function which only called slices.DeleteFunc
has been inlined.
