+++ internal/librarian/java/defaults.go
@@ -152,56 +152,6 @@ func tidyKeep(keep []string) []string {
 	return filteredKeepPaths
 }
 
-func isEmptyJavaAPI(j *config.JavaAPI) bool {
-	if j == nil {
-		return true
-	}
-	return !j.Monolithic &&
-		len(j.AdditionalProtos) == 0 &&
-		!j.OmitCommonResources &&
-		len(j.ExcludedProtos) == 0 &&
-		len(j.SkipProtoClassGeneration) == 0 &&
-		j.GAPICArtifactIDOverride == "" &&
-		j.GRPCArtifactIDOverride == "" &&
-		j.ProtoArtifactIDOverride == "" &&
-		j.GenerateGAPIC == nil &&
-		j.GenerateProto == nil &&
-		j.GenerateGRPC == nil &&
-		j.GenerateResourceNames == nil &&
-		len(j.CopyFiles) == 0 &&
-		j.Samples == nil
-}
-
-func isEmptyJavaModule(j *config.JavaModule) bool {
-	if j == nil {
-		return true
-	}
-	return j.APIIDOverride == "" &&
-		j.APIReference == "" &&
-		j.APIDescriptionOverride == "" &&
-		j.APIShortnameOverride == "" &&
-		j.ArtifactID == "" &&
-		j.ClientDocumentationOverride == "" &&
-		j.CodeownerTeam == "" &&
-		j.ExcludedDependencies == "" &&
-		len(j.ExcludedPOMs) == 0 &&
-		j.ExtraVersionedModules == "" &&
-		j.GroupID == "" &&
-		j.IssueTrackerOverride == "" &&
-		j.LibrariesBOMVersion == "" &&
-		j.LibraryTypeOverride == "" &&
-		j.MinJavaVersion == 0 &&
-		j.NamePrettyOverride == "" &&
-		j.ProductDocumentationOverride == "" &&
-		j.RecommendedPackage == "" &&
-		!j.BillingNotRequired &&
-		j.RestDocumentation == "" &&
-		j.RpcDocumentation == "" &&
-		j.TransportOverride == "" &&
-		!j.SkipPOMUpdates &&
-		!j.SkipAPIID
-}
-
 var (
 	// ErrOmitCommonResourcesConflict is returned when OmitCommonResources is true
 	// but common_resources.proto is also explicitly listed in AdditionalProtos.
