+++ internal/librarian/tidy_test.go
@@ -717,101 +717,6 @@ func TestTidy_DerivableRoots(t *testing.T) {
 	}
 }
 
-func TestTidyLanguageConfig_Rust(t *testing.T) {
-	for _, test := range []struct {
-		name        string
-		cfg         *config.Config
-		wantNumLibs int
-		wantNumMods int
-	}{
-		{
-			name: "empty_module_removed",
-			cfg: &config.Config{
-				Language: config.LanguageRust,
-				Sources: &config.Sources{
-					Googleapis: &config.Source{
-						Commit: "94ccedca05acb0bb60780789e93371c9e4100ddc",
-						SHA256: "fff40946e897d96bbdccd566cb993048a87029b7e08eacee3fe99eac792721ba",
-					},
-				},
-				Default: &config.Default{
-					Output: "generated/",
-				},
-				Libraries: []*config.Library{
-					{
-						Name:   "google-cloud-storage",
-						Output: "src/storage",
-						Rust: &config.RustCrate{
-							Modules: []*config.RustModule{
-								{
-									Output:   "src/storage/src/generated/protos/storage",
-									APIPath:  "google/storage/v2",
-									Template: "prost",
-								},
-								{
-									Output: "src/storage/control",
-								},
-							},
-						},
-					},
-				},
-			},
-			wantNumLibs: 1,
-			wantNumMods: 1, // Modules should be removed
-		},
-		{
-			name: "storage_module_not_removed",
-			cfg: &config.Config{
-				Language: config.LanguageRust,
-				Sources: &config.Sources{
-					Googleapis: &config.Source{
-						Commit: "94ccedca05acb0bb60780789e93371c9e4100ddc",
-						SHA256: "fff40946e897d96bbdccd566cb993048a87029b7e08eacee3fe99eac792721ba",
-					},
-				},
-				Default: &config.Default{
-					Output: "generated/",
-				},
-				Libraries: []*config.Library{
-					{
-						Name:   "google-cloud-storage",
-						Output: "src/storage",
-						Rust: &config.RustCrate{
-							Modules: []*config.RustModule{
-								{
-									Output:   "src/storage/src/generated/protos/storage",
-									Template: "storage",
-								},
-							},
-						},
-					},
-				},
-			},
-			wantNumLibs: 1,
-			wantNumMods: 1, // Rust storage module should NOT be removed
-		},
-	} {
-		t.Run(test.name, func(t *testing.T) {
-			tempDir := t.TempDir()
-
-			RunTidyOnConfig(t.Context(), tempDir, test.cfg)
-
-			cfg, err := yaml.Read[config.Config](filepath.Join(tempDir, config.LibrarianYAML))
-			if err != nil {
-				t.Fatal(err)
-			}
-
-			if len(cfg.Libraries) != test.wantNumLibs {
-				t.Fatalf("wrong number of libraries")
-			}
-			lib := cfg.Libraries[0]
-			if len(lib.Rust.Modules) != test.wantNumMods {
-				t.Fatalf("wrong number of modules")
-			}
-		})
-	}
-}
-
 func TestTidy_UnusedSections(t *testing.T) {
 	for _, test := range []struct {
 		name        string
